removed temporary files,  pep8 and updated documentation
