+++ docs/content/example_usage.rst
@@ -157,3 +157,34 @@ plot the counts using the `--log1p` option.
     :align: center
 
     Corrected Hi-C counts in log scale.
+
+
+TAD calling
+^^^^^^^^^^^
+
+To call TADs a corrected matrix is needed. Restriction fragment resolution matrices provide the best results.
+TAD calling works in two steps: First HiCExplorer computes a TAD-separation score based on a z-score matrix for
+all bins. Then those bins having a local minimum of the TAD-separation score are evaluated with respect to the
+surrounding bins to decide assign a p-value. Then a cutoff is applied to select the bins more likely to be TAD
+boundaries.
+
+.. code-block:: bash
+
+   $ hicFindTADs -m hic_corrected.npz --outPrefix hic_corrected --numberOfProcessors 16
+
+
+This code will produce several files: 1. The TAD-separation score file, 2. the z-score matrix, 3. a bed file
+with the boundary location, 4. a bed file with the domains, 5. a bedgraph file with the TAD-score that can be
+visualized in a genome browser.
+
+The TAD-separation score and the matrix can be visualized using :ref:`hicPlotTADs`.
+
+.. figure:: ../images/chorogenome_example.jpg
+    :scale: 90 %
+    :align: center
+
+    Example output from hicPlotTADs from http://chorogenome.ie-freiburg.mpg.de/
+
+
+
+
